Initiate testing with repository GET API call
The Storyline project must be built via TDD through and through. The
first step in this direction to establish a basic API for returning the
registered repositories (empty by default).

The app is now extracted into a it's own package as it must be a
reusable code for the testing and the application running too.