Add api call to find repos in directory
The application requires a way to specify a repository to work with. To
locate a repository through the GUI, the system must provide a way for
the UI code to check the directory structure on the computer and able to
tell if a particular directory is a Git repository or not.

The distinction between worktree's and bare repositories are subtle and
might not be needed much, but the repository is identified by the name
of the repository directory, but the parenting works differently for
bare repositories and normal checkouts.